+++ IOSMatchingEngineSDK/MobiledgeXiOSLibrary/Classes/MatchingEngine/QoSPositionKpi.swift
@@ -130,7 +130,7 @@ extension MobiledgeXiOSLibrary.MatchingEngine {
     ///   - port: port override of the dme server port
     ///   - request: QosKPIRequest dictionary, from createQosKPIRequest.
     /// - Returns: API Dictionary/json
-    public func getQosKPIPosition(host: String, port: UInt, request: [String: Any]) -> Promise<[String: AnyObject]>
+    public func getQosKPIPosition(host: String, port: UInt16, request: [String: Any]) -> Promise<[String: AnyObject]>
     {
         let promiseInputs: Promise<[String: AnyObject]> = Promise<[String: AnyObject]>.pending()
         os_log("getQosKPIPosition", log: OSLog.default, type: .debug)
